+++ test/initial_sync_repro_test.dart
@@ -30,7 +30,7 @@ class FakeGanttSyncClient implements GanttSyncClient {
   Future<String> getMerkleRoot() async => '';
 
   @override
-  Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+  Future<void> syncWithMerkle({required MerkleTree localTree}) async {}
 
   @override
   String get actorId => 'test-actor';