+++ test/src/legacy_gantt_view_model_test.dart
@@ -987,7 +987,7 @@ class MockGanttSyncClient extends GanttSyncClient {
   Future<String> getMerkleRoot() async => '';
 
   @override
-  Future<void> syncWithMerkle({required String remoteRoot, required int depth}) async {}
+  Future<void> syncWithMerkle({required MerkleTree localTree}) async {}
 
   @override
   String get actorId => 'mock-actor';